[chore] [pdata] Split internal/wrapper_common.go
To follow the established pattern with one struct per file
