+++ pdata/internal/wrapper_tracestate.go
@@ -1,0 +1,38 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package internal // import "go.opentelemetry.io/collector/pdata/internal"
+
+type TraceState struct {
+	orig *string
+}
+
+func GetOrigTraceState(ms TraceState) *string {
+	return ms.orig
+}
+
+func NewTraceState(orig *string) TraceState {
+	return TraceState{orig: orig}
+}
+
+func GenerateTestTraceState() TraceState {
+	var orig string
+	ms := NewTraceState(&orig)
+	FillTestTraceState(ms)
+	return ms
+}
+
+func FillTestTraceState(dest TraceState) {
+	*dest.orig = "rojo=00f067aa0ba902b7"
+}
